requirements.txt
aiohttp
Beaker
cairosvg
connexion==2.6.0
cssselect
Elasticsearch
Flask
Flask-Admin
Flask-APScheduler
Flask-Migrate
flask-jwt-extended
Flask-Cors
flask-restx
Flask-SQLAlchemy
Flask-Security
jsonschema
lxml
pandas
Petrovich
psycopg2
python-dateutil
requests
s3fs
tinycss
transliterate
Werkzeug
uWSGI
celery
redis
pyTelegramBotAPI
Pillow
email_validator
bcrypt
fuzzywuzzy
Jinja2
pyjwt
mimesis
gunicorn
python-dotenv
yookassa
SQLAlchemy-Utils